cleans up variable mutations and gets the first variable mutation test to pass

+++ app/src/main/java/little/horse/common/objects/metadata/WFRunVariableTypeEnum.java
@@ -4,7 +4,7 @@ public enum WFRunVariableTypeEnum {
     OBJECT,
     ARRAY,
     INT,
-    DOUBLE,
+    FLOAT,
     BOOLEAN,
     STRING;
 }
