+++ lhctl/lh_lib/schema/wf_spec_schema.py
@@ -20,7 +20,7 @@ class WFRunVariableTypeEnum(Enum):
     OBJECT = 'OBJECT'
     ARRAY = 'ARRAY'
     INT = 'INT'
-    DOUBLE = 'DOUBLE'
+    FLOAT = 'FLOAT'
     BOOLEAN = 'BOOLEAN'
     STRING = 'STRING'
